mlx5: DR, Expand DR STE callback and consolidate common structs
Enhance the Direct Rule (DR) STE mechanism by expanding the callback
structure to support the next-generation hardware (ConnectX-8).
This update provides greater flexibility in assigning values specific to
each hardware generation. Additionally, consolidate common enums and
structures into a shared header to improve code organization and
maintainability.
