+++ providers/mlx5/dr_ste.c
@@ -604,7 +604,7 @@ void dr_ste_set_actions_tx(struct dr_ste_ctx *ste_ctx,
 			   struct dr_ste_actions_attr *attr,
 			   uint32_t *added_stes)
 {
-	ste_ctx->set_actions_tx(action_type_set, ste_ctx->actions_caps,
+	ste_ctx->set_actions_tx(ste_ctx, action_type_set, ste_ctx->actions_caps,
 				hw_ste_arr, attr, added_stes);
 }
 
@@ -614,7 +614,7 @@ void dr_ste_set_actions_rx(struct dr_ste_ctx *ste_ctx,
 			   struct dr_ste_actions_attr *attr,
 			   uint32_t *added_stes)
 {
-	ste_ctx->set_actions_rx(action_type_set, ste_ctx->actions_caps,
+	ste_ctx->set_actions_rx(ste_ctx, action_type_set, ste_ctx->actions_caps,
 				hw_ste_arr, attr, added_stes);
 }
